+++ src/DataContainer/EntityImportSourceContainer.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2020 Heimrich & Hannot GmbH
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -10,7 +10,6 @@ namespace HeimrichHannot\EntityImportBundle\DataContainer;
 
 use Contao\Database;
 use Contao\DataContainer;
-use Contao\Message;
 use Contao\Model;
 use Contao\System;
 use HeimrichHannot\EntityImportBundle\Event\AddSourceFieldMappingPresetsEvent;
@@ -59,6 +58,10 @@ class EntityImportSourceContainer
     protected $database;
     protected $cache;
     /**
+     * @var EntityImportUtil
+     */
+    protected $entityImportUtil;
+    /**
      * @var FileUtil
      */
     private $fileUtil;
@@ -88,7 +91,7 @@ class EntityImportSourceContainer
      */
     private $databaseUtil;
 
-    public function __construct(SourceFactory $sourceFactory, FileUtil $fileUtil, ModelUtil $modelUtil, DcaUtil $dcaUtil, EntityImportUtil $util, EventDispatcherInterface $eventDispatcher, DatabaseUtil $databaseUtil)
+    public function __construct(SourceFactory $sourceFactory, FileUtil $fileUtil, ModelUtil $modelUtil, DcaUtil $dcaUtil, EntityImportUtil $util, EventDispatcherInterface $eventDispatcher, DatabaseUtil $databaseUtil, EntityImportUtil $entityImportUtil)
     {
         $this->activeBundles = System::getContainer()->getParameter('kernel.bundles');
         $this->sourceFactory = $sourceFactory;
@@ -98,6 +101,7 @@ class EntityImportSourceContainer
         $this->util = $util;
         $this->eventDispatcher = $eventDispatcher;
         $this->databaseUtil = $databaseUtil;
+        $this->entityImportUtil = $entityImportUtil;
     }
 
     public function setPreset(?DataContainer $dc)
@@ -136,6 +140,9 @@ class EntityImportSourceContainer
                             array_combine($options, $options)
                         );
                     } catch (\Exception $e) {
+                        $this->entityImportUtil->handleSourceConfigError(sprintf($GLOBALS['TL_LANG']['tl_entity_import_config']['error']['errorMessage'], $e->getMessage()));
+
+                        break;
                     }
                 }
 
@@ -150,7 +157,14 @@ class EntityImportSourceContainer
                         $source = $this->sourceFactory->createInstance($sourceModel->id);
 
                         $options = [];
-                        $fields = $source->getHeadingLine();
+
+                        try {
+                            $fields = $source->getHeadingLine();
+                        } catch (\Exception $e) {
+                            $this->entityImportUtil->handleSourceConfigError(sprintf($GLOBALS['TL_LANG']['tl_entity_import_config']['error']['errorMessage'], $e->getMessage()));
+
+                            break;
+                        }
 
                         foreach ($fields as $index => $field) {
                             if ($sourceModel->csvHeaderRow) {
@@ -177,7 +191,13 @@ class EntityImportSourceContainer
                         /** @var RSSFileSource $source */
                         $source = $this->sourceFactory->createInstance($sourceModel->id);
 
-                        $options = $source->getPostFieldsAsOptions();
+                        try {
+                            $options = $source->getPostFieldsAsOptions();
+                        } catch (\Exception $e) {
+                            $this->entityImportUtil->handleSourceConfigError($e->getMessage());
+
+                            break;
+                        }
 
                         $this->util->transformFieldMappingSourceValueToSelect(
                             array_combine($options, $options)
@@ -225,7 +245,13 @@ class EntityImportSourceContainer
             return '';
         }
 
-        return $this->getFileContent($sourceModel);
+        try {
+            $content = $this->getFileContent($sourceModel);
+        } catch (\Exception $e) {
+            return '';
+        }
+
+        return $content;
     }
 
     public function getFileContent(Model $sourceModel)
@@ -244,8 +270,6 @@ class EntityImportSourceContainer
 
             case static::FILETYPE_RSS:
                 return substr($source->getFileContent(true), 0, 50000);
-
-                break;
         }
 
         return $source->getFileContent(true);
@@ -260,7 +284,7 @@ class EntityImportSourceContainer
         try {
             $options = array_values(Database::getInstance($source->row())->listTables(null, true));
         } catch (\Exception $e) {
-            Message::addError(sprintf($GLOBALS['TL_LANG']['MSC']['entityImport']['dbConnectionError'], $e->getMessage()));
+            $this->entityImportUtil->handleSourceConfigError(sprintf($GLOBALS['TL_LANG']['MSC']['entityImport']['dbConnectionError'], $e->getMessage()));
 
             return [];
         }