+++ src/Source/RSSFileSource.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2020 Heimrich & Hannot GmbH
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -40,8 +40,6 @@ class RSSFileSource extends AbstractFileSource
         }
 
         $rss = new \DOMDocument();
-        $rss->loadXML($fileContent);
-
         $items = [];
 
         foreach ($rss->getElementsByTagName('item') as $key => $node) {